added sampleTime.
in most cases we check at what time the video player is up to,
and use the time for sampling buffers in the code below.
but if a sampleTime is provided then we use it instead of video player time.
sampleTime is used when synchronous time is needed (video player plays asynchronously),
such as when needing to access the video frames, frame by frame.
sampleTime is cleared (invalidated) on every frame so it must be set before calling update.